+++ psPAS/Functions/Accounts/Add-PASDependentAccount.ps1
@@ -45,12 +45,20 @@ function Add-PASDependentAccount {
 
         Assert-VersionRequirement -RequiredVersion 14.6
 
+        #The url pattern is different between self-hosted and ISPSS
+        #check for hosting type here
+        if (Test-IsISPSS) {
+            $URLString = 'account-dependents'
+        } else {
+            $URLString = 'dependentAccounts'
+        }
+
     }#begin
 
     process {
 
         #Create URL for Request
-        $URI = "$($psPASSession.BaseURI)/API/Accounts/$AccountID/dependentAccounts"
+        $URI = "$($psPASSession.BaseURI)/API/Accounts/$AccountID/$URLString"
 
         #Get all parameters that will be sent in the request
         $boundParameters = $PSBoundParameters | Get-PASParameter -ParametersToRemove AccountID